# perf: add microbenchmarks for a simple diamond with string manipulation