--- a/benches/Cargo.toml
+++ b/benches/Cargo.toml
@@ -15,6 +15,7 @@ workspace = true
 [dev-dependencies]
 criterion = { version = "0.5.0", features = [ "async_tokio", "html_reports" ] }
 hydroflow = { path = "../hydroflow", features = [ "debugging" ] }
+nameof = "1.0.0"
 rand = "0.8.0"
 rand_distr = "0.4.3"
 seq-macro = "0.2.0"
@@ -61,4 +62,8 @@ harness = false
 
 [[bench]]
 name = "symmetric_hash_join"
+harness = false
+
+[[bench]]
+name = "words_diamond"
 harness = false