--- a/benches/README.md
+++ b/benches/README.md
@@ -11,3 +11,5 @@ Run specific benchmarks:
 ```
 cargo bench -p benches --bench reachability
 ```
+
+Wordlist is from https://github.com/dwyl/english-words/blob/master/words_alpha.txt