… Cargo.lock
… benches/benches/words_alpha.txt
… benches/words_alpha.txt