`cargo bench -p benches --bench words_diamond`

```
hydroflow_diamond
                        time:   [72.142 ms 72.831 ms 73.535 ms]
hydroflo2_diamond_forloop
                        time:   [49.023 ms 49.128 ms 49.291 ms]
hydroflo2_diamond_iter_clone_chain
                        time:   [60.333 ms 60.675 ms 61.171 ms]
hydroflo2_diamond_iter_clone_interleave
                        time:   [67.000 ms 67.275 ms 67.613 ms]
hydroflo2_diamond_iter_buffer_chain
                        time:   [62.588 ms 63.017 ms 63.621 ms]
hydroflo2_diamond_iter_tee_chain
                        time:   [59.979 ms 60.390 ms 60.894 ms]
hydroflo2_diamond_iter_tee_interleave
                        time:   [62.646 ms 62.937 ms 63.242 ms]
hydroflo2_diamond_iter_buffer_one
                        time:   [58.574 ms 58.916 ms 59.327 ms]
```

![image](https://github.com/user-attachments/assets/20367b2e-728a-4e1c-99cf-a4d65a7228f8)